+++ src/ska_tmc_cdm/messages/subarray_node/configure/receptorgroup.py
@@ -16,7 +16,6 @@ class TrajectoryType(str, Enum):
     """
 
     FIXED = "fixed"
-    MOSAIC = "mosaic"
     SPIRAL = "spiral"
     RASTER = "raster"
     CONSTANT_VELOCITY = "constant-velocity"
@@ -24,36 +23,27 @@ class TrajectoryType(str, Enum):
     HYPOTROCHOID = "hypotrochoid"
 
 
-class MosaicTrajectory(CdmObject):
-    name: Literal[TrajectoryType.MOSAIC] = TrajectoryType.MOSAIC
-    attrs: MosaicTrajectory.Attrs
-
-    class Attrs(CdmObject):
-        x_offsets: list[float]
-        y_offsets: list[float]
-
-
 class TableTrajectory(CdmObject):
     name: Literal[TrajectoryType.TABLE] = TrajectoryType.TABLE
     attrs: TableTrajectory.Attrs
 
     class Attrs(CdmObject):
-        x: float
-        y: float
-        t: list[float] = Field(default_factory=list)
+        x: list[float]
+        y: list[float]
+        t: list[float]
 
 
-class FixedTrajectoryConfig(CdmObject):
+class FixedTrajectory(CdmObject):
     name: Literal[TrajectoryType.FIXED] = TrajectoryType.FIXED
+    attrs: FixedTrajectory.Attrs
 
     class Attrs(CdmObject):
         x: float
         y: float
-        t: list[float] = Field(default_factory=list)
 
 
 Trajectory = Annotated[
-    Union[MosaicTrajectory, TableTrajectory, FixedTrajectoryConfig],
+    Union[TableTrajectory, FixedTrajectory],
     Discriminator("name"),
 ]
 
@@ -78,11 +68,11 @@ class ProjectionAlignment(str, Enum):
 
 class Projection(CdmObject):
     """
-    Projection Config
+    Projection defines the projection for trajectory offsets.
     """
 
-    name: ProjectionType = ProjectionType.SIN
-    alignment: ProjectionAlignment = ProjectionAlignment.ICRS
+    name: Optional[ProjectionType] = ProjectionType.SSN
+    alignment: Optional[ProjectionAlignment] = ProjectionAlignment.ICRS
 
 
 class ReceptorGroup(CdmObject):
@@ -90,7 +80,7 @@ class ReceptorGroup(CdmObject):
     Receptor Group defines a set of receptors plus the sky direction and
     tracking/mapping strategy for target tracking."""
 
-    receptors: set[str] = Field(default_factory=set)
-    field: SkyDirection
+    receptors: Optional[set[str]] = Field(default_factory=set)
+    field: Optional[SkyDirection] = None
     trajectory: Optional[Trajectory] = None
     projection: Optional[Projection] = None